Refactor BlockPos construction to use consistent doubles

+++ src/main/kotlin/best/spaghetcodes/kira/utils/WorldUtils.kt
@@ -33,7 +33,11 @@ object WorldUtils {
         val lookVec = EntityUtils.get2dLookVec(player)
         for (i in 1..distance.toInt()) {
             val block = kira.mc.theWorld.getBlockState(
-                BlockPos(eyePos.x + lookVec.xCoord * i, eyePos.y.toDouble(), eyePos.z + lookVec.zCoord * i)
+                BlockPos(
+                    eyePos.x.toDouble() + lookVec.xCoord * i,
+                    eyePos.y.toDouble(),
+                    eyePos.z.toDouble() + lookVec.zCoord * i
+                )
             ).block
             if (block != Blocks.air) {
                 return true
